BN-1069 Move Bifrist configuration case classes to separated module, add separate config for network properties
