+++ node/src/main/resources/application.conf
@@ -22,7 +22,10 @@ bifrost {
     // A comma-delimited list of host:port pairs to connect to initially (i.e. `1.2.3.4:9085,5.6.7.8:9085`)
     known-peers = ""
     // Use actors system
-    // experimental = true
+    network-properties {
+      experimental = false
+      ping-pong-interval = 300 seconds
+    }
   }
   // Settings for RPC/gRPC
   rpc {
